Switched to non interleaved buffers

+++ JackedEngine/Rendering/Memory/AllocationFactories/BaseAllocationFactory.h
@@ -7,9 +7,9 @@
 
 class BaseAllocationFactory {
 public:
-	virtual const BaseVertexBufferAllocation* CreateVertexBufferAllocation(const void* data, const size_t dataSize) const = 0;
-	virtual const BaseIndexBufferAllocation* CreateIndexBufferAllocation(const void* data, const size_t dataSize) const = 0;
-	virtual const BaseUniformBufferAllocation* CreateUniformBufferAllocation(const size_t dataSize) const = 0;
+	virtual const BaseVertexBufferAllocation* CreateVertexBufferAllocation(const void* data, const uint32_t dataSize) const = 0;
+	virtual const BaseIndexBufferAllocation* CreateIndexBufferAllocation(const void* data, const uint32_t dataSize) const = 0;
+	virtual const BaseUniformBufferAllocation* CreateUniformBufferAllocation(const uint32_t dataSize) const = 0;
 	virtual const BaseImageBufferAllocation* CreateImageBufferAllocation(const void* data, const int width, const int height) const = 0;
 	virtual const BaseCubemapBufferAllocation* CreateCubemapBufferAllocation(const void* frontFaceData, const void* backFaceData, const void* upFaceData, const void* downFaceData, const void* rightFaceData, const void* leftFaceData, const int width, const int height) const = 0;
