+++ JackedEngine/Rendering/Memory/AllocationFactories/VMAAllocationFactory.cpp
@@ -6,15 +6,15 @@ VMAAllocationFactory::VMAAllocationFactory(const Device& device) :
 
 VMAAllocationFactory::~VMAAllocationFactory(){}
 
-const VMAVertexBufferAllocation* VMAAllocationFactory::CreateVertexBufferAllocation(const void* data, const size_t dataSize) const {
+const VMAVertexBufferAllocation* VMAAllocationFactory::CreateVertexBufferAllocation(const void* data, const uint32_t dataSize) const {
 	return new VMAVertexBufferAllocation(allocator, data, dataSize);
 }
 
-const VMAIndexBufferAllocation* VMAAllocationFactory::CreateIndexBufferAllocation(const void* data, const size_t dataSize) const {
+const VMAIndexBufferAllocation* VMAAllocationFactory::CreateIndexBufferAllocation(const void* data, const uint32_t dataSize) const {
 	return new VMAIndexBufferAllocation(allocator, data, dataSize);
 }
 
-const VMAUniformBufferAllocation* VMAAllocationFactory::CreateUniformBufferAllocation(const size_t bufferSize) const {
+const VMAUniformBufferAllocation* VMAAllocationFactory::CreateUniformBufferAllocation(const uint32_t bufferSize) const {
 	return new VMAUniformBufferAllocation(allocator, bufferSize);
 }
